feat(ts): stage 8 — migrate all tests to TypeScript
General changes across all test files:
- .test.js / .test.jsx → .test.ts / .test.tsx
- Test data typed with domain interfaces (Course, Author, UserState etc.)
- Typed preloaded state via TestPreloadedState interface
- vi.mocked() for type-safe access to mock methods
- buildWrapper typed with { children: ReactNode }
- Event objects cast with as React.ChangeEvent<HTMLInputElement>
- as unknown as React.FormEvent for synthetic form events

Key TypeScript patterns introduced:
- vi.mocked(fn) — tells TypeScript fn has mock methods
- as any — used for Axios response mocks (response shape not worth typing in tests)
- ReturnType<typeof buildStore> — inferred store type
- TestPreloadedState interface — ensures test data matches slice interfaces
