+++ src/components/Header/Header.test.tsx
@@ -1,21 +1,42 @@
-import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { configureStore } from '@reduxjs/toolkit';
 import { MemoryRouter as Router } from 'react-router-dom';
+import { configureStore } from '@reduxjs/toolkit';
 import userReducer from '../../store/user/reducer';
 import coursesReducer from '../../store/courses/reducer';
 import authorsReducer from '../../store/authors/reducer';
+import enrollmentsReducer from '../../store/enrollments/reducer';
 import Header from './Header';
 import { logoutUserService } from '../../services';
+import type { UserState, CoursesState, AuthorsState, EnrollmentsState } from '../../types';
 
 vi.mock('../../services');
 
-const initialState = {
+const mockNavigate = vi.fn();
+
+vi.mock('react-router-dom', async () => {
+  const actual = await vi.importActual('react-router-dom');
+  return {
+    ...actual,
+    useNavigate: () => mockNavigate,
+  };
+});
+
+// Typed preloaded state — TypeScript verifies shape matches slice interfaces.
+// This catches errors where test data is missing required fields.
+interface TestPreloadedState {
+  user: UserState;
+  courses: CoursesState;
+  authors: AuthorsState;
+  enrollments: EnrollmentsState;
+}
+
+const initialState: TestPreloadedState = {
   user: {
     isAuth: true,
     name: 'username',
     role: 'user',
+    email: 'user@test.com',
     status: 'succeeded',
     error: null,
   },
@@ -29,29 +50,25 @@ const initialState = {
     status: 'idle',
     error: null,
   },
+  enrollments: {
+    enrollments: [],
+    status: 'idle',
+    error: null,
+  },
 };
 
-const buildStore = (state = initialState) =>
+const buildStore = (state: TestPreloadedState = initialState) =>
   configureStore({
     reducer: {
       user: userReducer,
       courses: coursesReducer,
       authors: authorsReducer,
+      enrollments: enrollmentsReducer,
     },
     preloadedState: state,
   });
 
-const mockNavigate = vi.fn();
-
-vi.mock('react-router-dom', async () => {
-  const actual = await vi.importActual('react-router-dom');
-  return {
-    ...actual,
-    useNavigate: () => mockNavigate,
-  };
-});
-
-const renderHeader = (state = initialState) =>
+const renderHeader = (state: TestPreloadedState = initialState) =>
   render(
     <Provider store={buildStore(state)}>
       <Router>
@@ -63,7 +80,7 @@ const renderHeader = (state = initialState) =>
 describe('Header Component', () => {
   beforeEach(() => {
     vi.clearAllMocks();
-    logoutUserService.mockResolvedValue({ data: { successful: true } });
+    vi.mocked(logoutUserService).mockResolvedValue({ data: { successful: true } } as any);
   });
 
   it('should contain a logo', () => {
@@ -82,7 +99,7 @@ describe('Header Component', () => {
   });
 
   it('should display Login button when user is not authenticated', () => {
-    const guestState = {
+    const guestState: TestPreloadedState = {
       ...initialState,
       user: { ...initialState.user, isAuth: false, name: null },
     };
@@ -94,7 +111,7 @@ describe('Header Component', () => {
     renderHeader();
     fireEvent.click(screen.getByText('Logout'));
     await waitFor(() => {
-      expect(logoutUserService).toHaveBeenCalledTimes(1);
+      expect(vi.mocked(logoutUserService)).toHaveBeenCalledTimes(1);
       expect(mockNavigate).toHaveBeenCalledWith('/login');
     });
   });
